refactor: Extract config-utils from config-types
...to avoid module import problems when generating JSON Schemas off the config-types file.
