+++ app/configurator/components/badges.tsx
@@ -1,6 +1,6 @@
 import { Badge, BadgeProps } from "@mui/material";
 
-import { getChartConfig, useChartConfigFilters } from "@/configurator";
+import { getChartConfig, useChartConfigFilters } from "@/config-utils";
 import { useControlSectionContext } from "@/configurator/components/chart-controls/section";
 import {
   isConfiguring,
