+++ app/configurator/components/chart-options-selector.tsx
@@ -32,7 +32,6 @@ import {
 import { InfoIconTooltip } from "@/components/info-icon-tooltip";
 import { MaybeTooltip } from "@/components/maybe-tooltip";
 import { TooltipTitle } from "@/components/tooltip-utils";
-import { GenericField } from "@/config-types";
 import {
   AnimationField,
   ChartConfig,
@@ -44,17 +43,16 @@ import {
   ComboLineSingleConfig,
   ComponentType,
   ConfiguratorStateConfiguringChart,
-  getChartConfig,
+  GenericField,
   ImputationType,
   imputationTypes,
   isAnimationInConfig,
   isComboChartConfig,
-  isConfiguring,
   isTableConfig,
   MapConfig,
   SortingType,
-  useConfiguratorState,
-} from "@/configurator";
+} from "@/config-types";
+import { getChartConfig } from "@/config-utils";
 import { ColorPalette } from "@/configurator/components/chart-controls/color-palette";
 import { ColorRampField } from "@/configurator/components/chart-controls/color-ramp";
 import {
@@ -82,6 +80,10 @@ import {
   canUseAbbreviations,
   getComponentLabel,
 } from "@/configurator/components/ui-helpers";
+import {
+  isConfiguring,
+  useConfiguratorState,
+} from "@/configurator/configurator-state";
 import { FIELD_VALUE_NONE } from "@/configurator/constants";
 import { TableColumnOptions } from "@/configurator/table/table-chart-options";
 import {
