+++ app/configurator/components/configurator.tsx
@@ -31,15 +31,9 @@ import {
   ConfiguratorState,
   ConfiguratorStateLayouting,
   enableLayouting,
-  getChartConfig,
-  hasChartConfigs,
-  initChartStateFromChartEdit,
-  isConfiguring,
-  isLayouting,
   MetaKey,
-  saveChartLocally,
-  useConfiguratorState,
-} from "@/configurator";
+} from "@/config-types";
+import { getChartConfig } from "@/config-utils";
 import {
   ChartAnnotationsSelector,
   LayoutAnnotationsSelector,
@@ -66,6 +60,14 @@ import {
   usePreviewBreakpoint,
 } from "@/configurator/components/preview-breakpoint";
 import { ShowDrawerButton } from "@/configurator/components/show-drawer-button";
+import {
+  hasChartConfigs,
+  initChartStateFromChartEdit,
+  isConfiguring,
+  isLayouting,
+  saveChartLocally,
+  useConfiguratorState,
+} from "@/configurator/configurator-state";
 import { ChartConfiguratorTable } from "@/configurator/table/table-chart-configurator";
 import { useUserConfig } from "@/domain/user-configs";
 import { useDataCubesComponentsQuery } from "@/graphql/hooks";
