+++ app/pages/v/[chartId].tsx
@@ -22,7 +22,8 @@ import { ChartPublished } from "@/components/chart-published";
 import { PublishSuccess } from "@/components/hint";
 import { ContentLayout } from "@/components/layout";
 import { PublishActions } from "@/components/publish-actions";
-import { ConfiguratorStatePublished, getChartConfig } from "@/config-types";
+import { ConfiguratorStatePublished } from "@/config-types";
+import { getChartConfig } from "@/config-utils";
 import { ConfiguratorStateProvider } from "@/configurator/configurator-state";
 import { getConfig, increaseConfigViewCount } from "@/db/config";
 import { deserializeProps, Serialized, serializeProps } from "@/db/serialize";
